sm-rl- add put and get endpoints and tests

+++ src/main/java/edu/ucsb/cs156/happiercows/entities/UserCommons.java
@@ -26,7 +26,7 @@ public class UserCommons {
   @Column(name="user_id")
   private long userId;  
 
-  private int totalWealth;
+  private double totalWealth;
 
   private int numOfCows;
 }
